requirements.txt
minio==7.1.17
python-dotenv==1.0.0
